incorporate 32-bit platform fix from scipy/scipy#11229

+++ fast_upfirdn/cpu/_upfirdn.py
@@ -99,7 +99,9 @@ class _UpFIRDn(object):
         """Apply the prepared filter to the specified axis of a nD signal x"""
         if axis < -x.ndim or axis >= x.ndim:
             raise ValueError("axis out of range")
-        output_shape = np.asarray(x.shape)
+        # Explicit use of np.int64 for output_shape dtype avoids OverflowError
+        # when allocating large array on platforms where np.int_ is 32 bits
+        output_shape = np.asarray(x.shape, dtype=np.int64)
         if crop:
             # only compute output over roughly the original signal extent
             output_len = int(ceil(x.shape[axis] * self._up / self._down))
